refactor: `isMouseOverBenefitsWrapper` 모듈로 변경
